Move zellij helper actions to cmd/action subfolder
The action_enums.go and action_modes.go files in cmd/ contained only
helper functions with no init(), breaking the release optimizer which
expects every cmd/ file to be a subcommand with an init(). Moved these
to the cmd/action/ subfolder as exported functions and updated all
call sites to use the action package.
