+++ completers/common/zellij_completer/cmd/action_edit.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
+	"github.com/carapace-sh/carapace-bin/completers/common/zellij_completer/cmd/action"
 	"github.com/spf13/cobra"
 )
 
@@ -35,7 +36,7 @@ func init() {
 	carapace.Gen(action_editCmd).FlagCompletion(carapace.ActionMap{
 		"borderless": carapace.ActionValues("true", "false"),
 		"cwd":        carapace.ActionFiles(),
-		"direction":  actionDirections(),
+		"direction":  action.ActionDirections(),
 		"pinned":     carapace.ActionValues("true", "false"),
 	})
