--- completers/common/zellij_completer/cmd/action_enums.go
@@ -1,11 +1,0 @@
-package cmd
-
-import "github.com/carapace-sh/carapace"
-
-func actionDirections() carapace.Action {
-	return carapace.ActionValues("left", "down", "up", "right")
-}
-
-func actionResizes() carapace.Action {
-	return carapace.ActionValues("increase", "decrease")
-}
