+++ completers/common/zellij_completer/cmd/attach_options.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
+	"github.com/carapace-sh/carapace-bin/completers/common/zellij_completer/cmd/action"
 	"github.com/spf13/cobra"
 )
 
@@ -76,7 +77,7 @@ func init() {
 		"dangerously-enable-paste-buffer-read": carapace.ActionValues("true", "false"),
 		"default-cwd":                          carapace.ActionFiles(),
 		"default-layout":                       carapace.ActionFiles(),
-		"default-mode":                         actionModes(),
+		"default-mode":                         action.ActionModes(),
 		"default-shell":                        carapace.ActionFiles(),
 		"disable-session-metadata":             carapace.ActionValues("true", "false"),
 		"focus-follows-mouse":                  carapace.ActionValues("true", "false"),
